GlslEditor: Fix crash in function hint proposal
The function hint proposal gets filled with pointers
it does not own. It does not notice when these pointers
will get deleted or modified, but as long the
hint is displayed it tries to access them.
If the document gets modified and the gls parser tries
to update the document's AST these pointers might vanish.
Only use the needed information on the proposal's side
to avoid null pointer access.

Trade off: updating the proposal hint of declarations
           and definitions when document changes does
           not work on the fly

Task-number: QTCREATORBUG-20570
Change-Id: I20253e92e1ca9822b1fc1f327c5c5e16b4fd91be
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@qt.io>