Git: Abort mergetool if not configured
When merge.tool is not configured, mergetool is not started, and the
following message is displayed:

This message is displayed because 'merge.tool' is not configured.
See 'git mergetool --tool-help' or 'git help config' for more details.
'git mergetool' will now attempt to use one of the following tools:
opendiff kdiff3 tkdiff xxdiff meld tortoisemerge gvimdiff diffuse diffmerge
ecmerge p4merge araxis bc codecompare emerge vimdiff
Merging:
test.cpp

Normal merge conflict for 'test.cpp':
  {local}: modified file
  {remote}: modified file
Hit return to start merge resolution tool (vimdiff):

On Windows, vimdiff is the default, and since it is textual, if we use it,
we have no way to pass any input for it, then we reach a dead-end.

On Linux/KDE (if DISPLAY is set), Git defaults to kdiff3. This is better,
but we cannot count on it even being installed.

Just abort with a warning on this case, and let the user [install and]
configure the preferred (graphical) mergetool.

Change-Id: I7144d9348a4d1ee03176b71b4ac2242b16795403
Reviewed-by: André Hartmann <aha_1980@gmx.de>