Fix crash of process steps in deploy configuration
If the step is in a deploy configuration, it doesn't have a direct
parent build configuration, so we need to get that from the active
target.
This is generally improved for BuildStep in master, but here we still
have to do that by hand.

Fix-up of 75afda1798ebea05be9be8c4c415dd910a8966c9

Task-number: QTCREATORBUG-20631
Change-Id: I22fa16d424ee5928c920663e9d01c791b964e873
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>
Reviewed-by: Ivan Donchevskii <ivan.donchevskii@qt.io>