WinDebugInterface: Remove blocking connection to main thread
The debugOutput() signal was connected via BlockingQueuedConnection and
blocks the WinDebugInterface thread until the signal is received. If we
signal "buffer ready" afterwards, all OutputDebugString calls (of all
processes of the system) that wait for "buffer ready" also wait for our
internal signal delivery. This can lead to hangs in circumstances where
the event loop is currently unable to deliver the signal.

Remove the blocking queued connection, and introduce an internal
_q_debugOutputReady() signal that is emitted whenever new debug output
is available. This signal is queued-connected to a dispatchDebugOutput
function, which is running in the main thread. That function retrieves
the data from the WinDebugInterfaceThread and emits debugOutput(qint64,
QString). This massively reduces the event queue load without blocking
the WinDebugInterface thread.

Task-number: QTCREATORBUG-20640
Change-Id: I91f8f794af8da2a695c2b897f678844b142a5991
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>