Fix inserting tasks into issues pane
When tasks are inserted into the source model, the source indices in the
filtered model after the new items where not updated correctly.
The indices must be increased by the number of items added to the source
model, not the number of items added to the filtered model (which can be
less).

Task-number: QTCREATORBUG-20542
Task-number: QTCREATORBUG-18951
Change-Id: Idae9cf4241c31229dadf5c9fea383aef3fdfffb0
Reviewed-by: Orgad Shaneh <orgads@gmail.com>