QmlDesigner: Reparent only on CTRL+SHIFT
Shift is a common modifier and we should use it
to limit changes to one axis (e.g. x), since this
is a useful and typical usage.

We moved reparenting to CTRL+SHIFT. It has issues
and can trigger crashes. Still it can useful
and it is unlikely to be triggered randomly/by accident
on this shortcut.

Task-number: QDS-7
Change-Id: Ieeac6323659d5bd95bd27a05b8e18808ecdf390b
Reviewed-by: Brook Cronin <brook.cronin@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>