QmlDesigner: Fix crash when opening qmlproject while Design mode is open
Fix the order of signal handling, which was changed in
a5935cb27a0609764586302bf109796d74060f3d
but leads to this crash.
The connection to the DesignMode instance still needs to be
delayed to after it has been created in Core plugin's
extensionsInitialized, so do the final initialization as a reaction to
ICore::coreAboutToOpen.

Task-number: QTCREATORBUG-20495
Change-Id: I92d9a4ff689d2f191d2fd368966ca582bf8af2f7
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>