ProjectExplorer: only add Tasks from the main thread
If something fails while extracting the build env from msvc vavars
scripts we want to add task hub entries to inform the user about a most
probably unusable toolchain. As the msvc toolchain detection is threaded
nowadays and the TaskHub is not thread safe we need to propagate the
error message to the main thread before adding the Task entries.

Change-Id: I5c67d3b8d58b22ea895afb6f22ee73f19fa52f17
Reviewed-by: Eike Ziller <eike.ziller@qt.io>