Fix assert when filtering examples / tutorials
The examples are put throw two proxy models. The first one implements
the sorting and filtering, the second one transforms the list into a
grid.

The grid proxy model was based on QIdentityProxyModel, which is wrong
because that explicitly states that it should be used when the structure
of the model stays the same.
Even QAbstractProxyModel works under the assumption that all items in
the proxy model correspond to an item in the source model, which is not
true for a N*M grid.

Just implement the proxy model "by hand", making sure that all items
that cannot be mapped to the source are handles correctly as "empty"
items.

Task-number: QTCREATORBUG-20641
Change-Id: Id08b77bb7157b893b9d0d73dc006db98d90032b7
Reviewed-by: hjk <hjk@qt.io>