Revert "Wizards: Drop support for Qt4 for widgets application"
This reverts commit 1a6522b47c7eb667d297c71a48751e22f7069c04
which leaves people wondering about their kit setups.

Creating a project that possibly does not compile due to the
use of C++11's nullptr in the generated code is lesser evil.

Apart from that this potential mis-compilation does not
happen in practice on e.g. Ubuntu 18.04 with system Qt 4
and system compiler.

Change-Id: Iaa90f225d5317dc48428aa2d3fcf3ec051ef2018
Reviewed-by: Christian Stenger <christian.stenger@qt.io>