Improve macOS dark mode support
The current theme support in Creator works well, also
in macOS dark mode. This change makes sure that the
native macOS appearance setting (for the Creator process)
and the creator theme setting are in sync.

This is required to avoid light/dark palette issues for
the light Creator themes when the desktop is configured
for dark appearance.

Add a “DarkUserInterface” flag to the creatortheme files.
This flag indicates if the theme is overall dark, with
light text on dark background.

Make setCreatorTheme() force the standard light aqua
appearance for light Creator themes.

Change-Id: I1a5c183230ab0e66641fd834df19d7e0ad1b6a53
Reviewed-by: Eike Ziller <eike.ziller@qt.io>