Fix handling of read-only editors when applying refactorings
If e.g. a ".ui" file is open, there is a read-only text editor widget
for the file, even though the file itself is writable.
The application of refactorings or global text-based replace should not
change the content of this read-only editor widget, but instead operate
directly on the file as if it wasn't open in Qt Creator at all.
It should also silently reload these files after modification on disk.

Task-number: QTCREATORBUG-19958
Change-Id: I409d5d03059be4c3520a1031ff0fbfa9feb675bb
Reviewed-by: David Schulz <david.schulz@qt.io>