Doc: Use defines to show and hide info about Edit mode
Move the descriptions of sidebar views to
creator-sidebar-views.qdocinc to be able to include them
in the "Working in Edit Mode" topic in the Qt Design Studio
Manual.

Restructure the section and update the screenshot. A different
screenshot for the Qt Design Studio Manual will be added in
its doc sources.

Change-Id: Icd33c580c040ffefa5563fea723bd02e2d2ce520
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>