Fix editing of custom executable path
When editing the path for a custom executable run configuration, the
text cursor would jump to the end every time anything is typed. This
makes changing a part inside the text very cumbersome.

This happens because the executable aspect registers a "display filter"
that transforms the input to native separators.

Solve that issue generically for the path chooser by resetting its text
cursor position after the path has been set, if the input field has
focus.

Change-Id: Ic0a178e942da8df1e53b5d90c78a5bf1675865c2
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>