Tracing: Properly signal destruction of non-owned properties
When the zoomer, model, or notes properties for TimelineAbstractRenderer
are destroyed, the QPointers become null, and therefore the change
signals need to be emitted, and an update to the widget should be
scheduled.

If this is not done, clients using the property might keep a dangling
pointer in some cache and dereference it later.

Change-Id: I63fb657afb8138e9ac54c7d839d4a52271f8f3e8
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>