Theming: Fix value of TextColorDisabled for classic mode
TextColorDisabled has always been the same value as TextColorNormal
(which was presumably unintended), and nobody seems to have noticed for
a long time. However, as of commit f7a0a3cd26 we have been using
TextColorDisabled to display disabled project nodes, which means users
could not tell enabled and disabled parts of the project apart anymore
when using classic mode.
We now use the same value as in the other themes.

Change-Id: I1f616b6808d6c2012748af24b90953e4eb75df18
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>