CodePaster: Add support for pastecode.xyz
Our users are running out of choices for code pasting: KDE has required
an account for quite a while now, and pastebin.ca appears to be offline,
leaving them only with the ad-heavy pastebin.com.
Therefore, we add support for a new site: pastecode.xyz seems
responsive enough, has a simple API and has been around for a bit.

[ChangeLog] Added support for the pastecode.xyz code pasting service

Change-Id: I9b622853df24c28d65c6e388f953c82fc5380fae
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>