winrt: Use full path when running windeployqt
As winrt does not have the concept of a run environment acquiring the
path to windeployqt from the environment might lead to unexpected results
like using windeployqt from a different Qt installation. Calling the
program with its absolute path is the safer alternative and thus should
be the default anyways.

Task-number: QTBUG-70175
Change-Id: I6eb5416275d7ff243f396bf1fccdbcb9d04f7e15
Reviewed-by: Christian Stenger <christian.stenger@qt.io>