macOS: Fix passing environment to terminal
We do not actually start a new Terminal process, so we need to clean up
the environment after the fact.

Clean the environment in the started shell except for some essentials,
read the config files as if for a login shell, and re-export the
environment.

There still can be differences, since environment variables set in the
user's bash profile etc will "win". This is wrong if we want to open the
terminal in anything but the "system environment", especially if the
PATH is effected, but I don't see how to solve that without severely
crippling the shell setup. This is also the current state on Linux.

Change-Id: I1d3c8184ac3bf543675e96f73253085fa6b1b29d
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>