Squish: Wait for a clean shutdown when restarting QC
This fixes a race condition in tst_CSUP06 where the next run
of QC removed .user* but the first instance was not completely
finished with the shutdown process. The .user files of the project
got updated (or recreated) before the second instance tried to
open the same project again. Ensure the first instance is closed
to be sure that project's .user files got created or updated
before the second instance tries to remove and re-open them.
Follow the same approach for other tests to avoid the same issue
later on.

Change-Id: I37721f4dd647f9bbf7c6fed6e753a2906e30db81
Reviewed-by: Robert Loehning <robert.loehning@qt.io>